Moving Initializer into util.py
