+++ pyomo/core/base/set.py
@@ -8,7 +8,6 @@
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 import collections
-import functools
 import inspect
 import itertools
 import logging
@@ -26,13 +25,6 @@ except ImportError:
             ans -= b
         return ans
 
-if six.PY2:
-    getargspec = inspect.getargspec
-else:
-    # For our needs, getfullargspec is a drop-in replacement for
-    # getargspec (which was removed in Python 3.x)
-    getargspec = inspect.getfullargspec
-
 from six import iteritems, iterkeys
 from six.moves import xrange
 
@@ -41,6 +33,10 @@ from pyutilib.misc.misc import flatten_tuple
 from pyomo.common.deprecation import deprecated, deprecation_warning
 from pyomo.common.errors import DeveloperError
 from pyomo.common.timing import ConstructionTimer
+from pyomo.core.base.util import (
+    disable_methods, InitializerBase, Initializer, ConstantInitializer,
+    CountedCallInitializer, ItemInitializer, IndexedCallInitializer,
+)
 from pyomo.core.expr.numvalue import (
     native_types, native_numeric_types, as_numeric, value,
 )
@@ -55,363 +51,6 @@ logger = logging.getLogger('pyomo.core')
 _prePython37 = sys.version_info[:2] < (3,7)
 
 FLATTEN_CROSS_PRODUCT = True
-
-#
-# The following decorator is general and should probably be promoted to
-# component.py so that we can efficiently handle construction errors on
-# scalar components.
-#
-def _disable_method(fcn, msg=None):
-    if msg is None:
-        msg = 'access %s on' % (fcn.__func__.func_name,)
-    def impl(self, *args, **kwds):
-        raise RuntimeError(
-            "Cannot %s %s '%s' before it has been constructed (initialized)."
-            % (msg, type(self).__name__, self.name))
-
-    # functools.wraps doesn't preserve the function signature until
-    # Python 3.4.  For backwards compatability with Python 2.x, we will
-    # create a temporary (lambda) function using eval that matches the
-    # function signature passed in and calls the generic impl() function
-    args = inspect.formatargspec(*inspect.getargspec(fcn))
-    impl_args = eval('lambda %s: impl%s' % (args[1:-1], args), {'impl': impl})
-    return functools.wraps(fcn)(impl_args)
-
-def _disable_property(fcn, msg=None):
-    if msg is None:
-        _gmsg = 'access property %s on' % (fcn.fget.__name__,)
-    else:
-        _gmsg = msg
-    def getter(self, *args, **kwds):
-        raise RuntimeError(
-            "Cannot %s %s '%s' before it has been constructed (initialized)."
-            % (_gmsg, type(self).__name__, self.name))
-
-    if msg is None:
-        _smsg = 'set property %s on' % (fcn.fget.__name__,)
-    else:
-        _smsg = msg
-    def setter(self, *args, **kwds):
-        raise RuntimeError(
-            "Cannot %s %s '%s' before it has been constructed (initialized)."
-            % (_smsg, type(self).__name__, self.name))
-
-    return property(fget=getter, fset=setter, doc=fcn.__doc__)
-
-def disable_methods(methods):
-    """Class decorator to disable methods before construct is called.
-
-    This decorator should be called to create "Abstract" scalar classes
-    that override key methods to raise exceptions.  When the construct()
-    method is called, the class instance changes type back to the
-    original scalar component and the full class functionality is
-    restored.  The prevents most class methods from having to begin with
-    "`if not self.parent_component()._constructed: raise RuntimeError`"
-    """
-    def class_decorator(cls):
-        assert(len(cls.__bases__) == 1)
-        base = cls.__bases__[0]
-
-        def construct(self, data=None):
-            self.__class__ = base
-            return base.construct(self, data)
-        construct.__doc__ = base.construct.__doc__
-        cls.construct = construct
-
-        for method in methods:
-            if type(method) is tuple:
-                method, msg = method
-            else:
-                msg = None
-            if not hasattr(base, method):
-                raise DeveloperError(
-                    "Cannot disable method %s on %s: not present on base class"
-                    % (method, cls))
-            base_method = getattr(base, method)
-            if type(base_method) is property:
-                setattr(cls, method, _disable_property(base_method, msg))
-            else:
-                setattr(cls, method, _disable_method(base_method, msg))
-        return cls
-
-    return class_decorator
-
-
-#
-# The following set of "Initializer" classes are a general functionality
-# and should be promoted to their own module so that we can use them on
-# all Components to standardize how we process component arguments.
-#
-def Initializer(init,
-                allow_generators=False,
-                treat_sequences_as_mappings=True,
-                arg_not_specified=None):
-    if init.__class__ in native_types:
-        if init is arg_not_specified:
-            return None
-        return _ConstantInitializer(init)
-    elif inspect.isfunction(init):
-        if not allow_generators and inspect.isgeneratorfunction(init):
-            raise ValueError("Generator functions are not allowed")
-        # Historically pyomo.core.base.misc.apply_indexed_rule
-        # accepted rules that took only the parent block (even for
-        # indexed components).  We will preserve that functionality
-        # here.
-        _args = getargspec(init)
-        if len(_args.args) == 1 and _args.varargs is None:
-            return _ScalarCallInitializer(init)
-        else:
-            return _IndexedCallInitializer(init)
-    elif isinstance(init, collections.Mapping):
-        return _ItemInitializer(init)
-    elif isinstance(init, collections.Sequence) \
-            and not isinstance(init, six.string_types):
-        if treat_sequences_as_mappings:
-            return _ItemInitializer(init)
-        else:
-            return _ConstantInitializer(init)
-    elif inspect.isgenerator(init) or hasattr(init, 'next') \
-         or hasattr(init, '__next__'):
-        if not allow_generators:
-            raise ValueError("Generators are not allowed")
-        return _ConstantInitializer(init)
-    else:
-        return _ConstantInitializer(init)
-
-class _InitializerBase(object):
-    __slots__ = ()
-
-    verified = False
-
-    def __getstate__(self):
-        return dict((k, getattr(self,k)) for k in self.__slots__)
-
-    def __setstate__(self, state):
-        for key, val in iteritems(state):
-            object.__setattr__(self, key, val)
-
-class _ConstantInitializer(_InitializerBase):
-    __slots__ = ('val','verified')
-
-    def __init__(self, val):
-        self.val = val
-        self.verified = False
-
-    def __call__(self, parent, idx):
-        return self.val
-
-    def constant(self):
-        return True
-
-class _ItemInitializer(_InitializerBase):
-    __slots__ = ('_dict',)
-
-    def __init__(self, _dict):
-        self._dict = _dict
-
-    def __call__(self, parent, idx):
-        return self._dict[idx]
-
-    def constant(self):
-        return False
-
-class _IndexedCallInitializer(_InitializerBase):
-    __slots__ = ('_fcn',)
-
-    def __init__(self, _fcn):
-        self._fcn = _fcn
-
-    def __call__(self, parent, idx):
-        # Note: this is called by a component using data from a Set (so
-        # any tuple-like type should have already been checked and
-        # converted to a tuple; or flattening is turned off and it is
-        # the user's responsibility to sort things out.
-        if idx.__class__ is tuple:
-            return self._fcn(parent, *idx)
-        else:
-            return self._fcn(parent, idx)
-
-    def constant(self):
-        return False
-
-
-class _CountedCallGenerator(object):
-    def __init__(self, fcn, scalar, parent, idx):
-        # Note: this is called by a component using data from a Set (so
-        # any tuple-like type should have already been checked and
-        # converted to a tuple; or flattening is turned off and it is
-        # the user's responsibility to sort things out.
-        self._count = 0
-        if scalar:
-            self._fcn = lambda c: self._filter(fcn(parent, c))
-        elif idx.__class__ is tuple:
-            self._fcn = lambda c: self._filter(fcn(parent, c, *idx))
-        else:
-            self._fcn = lambda c: self._filter(fcn(parent, c, idx))
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        self._count += 1
-        return self._fcn(self._count)
-
-    @staticmethod
-    def _filter(x):
-        if x is None:
-            raise ValueError(
-                """Counted Set rule returned None instead of Set.End.
-    Counted Set rules of the form fcn(model, count, *idx) will be called
-    repeatedly with an increasing count parameter until the rule returns
-    Set.End.  None is not a valid Set member in this case due to the
-    likelihood that an error in the rule can incorrectly return None.""")
-        return x
-
-
-    next = __next__
-
-
-class _CountedCallInitializer(_InitializerBase):
-    # Pyomo has a historical feature for some rules, where the number of
-    # times[*1] the rule was called could be passed as an additional
-    # argument between the block and the index.  This was primarily
-    # supported by Set and ConstraintList.  There were many issues with
-    # the syntax, including inconsistent support for jagged (dimen=None)
-    # indexing sets, inconsistent support for *args rules, and a likely
-    # infinite loop if the rule returned Constraint.Skip.
-    #
-    # As a slight departure from previous implementations, we will ONLY
-    # allow the counted rule syntax when the rule does NOT use *args
-    #
-    # [*1] The extra argument was one-based, and was only incremented
-    #     when a valid value was returned by the rule and added to the
-    #     _data.  This was fragile, as returning something like
-    #     {Component}.Skip could result in an infinite loop.  This
-    #     implementation deviates from that behavior and increments the
-    #     counter every time the rule is called.
-    #
-    # [JDS 6/2019] We will support a slightly restricted but more
-    # consistent form of the original implementation for backwards
-    # compatability, but I believe that we should deprecate this syntax
-    # entirely.
-    __slots__ = ('_fcn','_is_counted_rule', '_scalar',)
-
-    def __init__(self, obj, _indexed_init):
-        self._fcn = _indexed_init._fcn
-        self._is_counted_rule = None
-        self._scalar = not obj.is_indexed()
-        if self._scalar:
-            self._is_counted_rule = True
-
-    def __call__(self, parent, idx):
-        # Note: this is called by a component using data from a Set (so
-        # any tuple-like type should have already been checked and
-        # converted to a tuple; or flattening is turned off and it is
-        # the user's responsibility to sort things out.
-        if self._is_counted_rule == False:
-            if idx.__class__ is tuple:
-                return self._fcn(parent, *idx)
-            else:
-                return self._fcn(parent, idx)
-        if self._is_counted_rule == True:
-            return _CountedCallGenerator(self._fcn, self._scalar, parent, idx)
-
-        # Note that this code will only be called once, and only if
-        # the object is not a scalar.
-        _args = getargspec(self._fcn)
-        _len = len(idx) if idx.__class__ is tuple else 1
-        if _len + 2 == len(_args.args):
-            self._is_counted_rule = True
-        else:
-            self._is_counted_rule = False
-        return self.__call__(parent, idx)
-
-    def constant(self):
-        return False
-
-class _ScalarCallInitializer(_InitializerBase):
-    __slots__ = ('_fcn',)
-
-    def __init__(self, _fcn):
-        self._fcn = _fcn
-
-    def __call__(self, parent, idx):
-        return self._fcn(parent)
-
-    def constant(self):
-        return False
-
-class SetInitializer(_InitializerBase):
-    __slots__ = ('_set','verified')
-
-    def __init__(self, init, allow_generators=True):
-        self.verified = False
-        if init is None:
-            self._set = None
-        else:
-            self._set = Initializer(
-                init, allow_generators=allow_generators,
-                treat_sequences_as_mappings=False)
-
-    def intersect(self, other):
-        if self._set is None:
-            if type(other) is SetInitializer:
-                self._set = other._set
-            else:
-                self._set = other
-        elif type(other) is SetInitializer:
-            if other._set is not None:
-                self._set = _SetIntersectInitializer(self._set, other._set)
-        else:
-            self._set = _SetIntersectInitializer(self._set, other)
-
-    def __call__(self, parent, idx):
-        if self._set is None:
-            return Any
-        else:
-            return self._set(parent, idx)
-
-    def constant(self):
-        return self._set is None or self._set.constant()
-
-    def setdefault(self, val):
-        if self._set is None:
-            self._set = _ConstantInitializer(val)
-
-class _SetIntersectInitializer(_InitializerBase):
-    __slots__ = ('_A','_B',)
-    def __init__(self, setA, setB):
-        self._A = setA
-        self._B = setB
-
-    def __call__(self, parent, idx):
-        return SetIntersection(self._A(parent, idx), self._B(parent, idx))
-
-    def constant(self):
-        return self._A.constant() and self._B.constant()
-
-class RangeSetInitializer(_InitializerBase):
-    __slots__ = ('_init', 'default_step',)
-    def __init__(self, init, default_step=1):
-        self._init = Initializer(init, treat_sequences_as_mappings=False)
-        self.default_step = default_step
-
-    def __call__(self, parent, idx):
-        val = self._init(parent, idx)
-        if not isinstance(val, collections.Sequence):
-            val = (1, val, self.default_step)
-        if len(val) < 3:
-            val = tuple(val) + (self.default_step,)
-        ans = RangeSet(*tuple(val))
-        ans.construct()
-        return ans
-
-    def constant(self):
-        return self._init.constant()
-
-    def setdefault(self, val):
-        # This is a real range set... there is no default to set
-        pass
 
 
 def process_setarg(arg):
@@ -517,6 +156,78 @@ def simple_set_rule( fn ):
 class RangeDifferenceError(ValueError): pass
 
 class UnknownSetDimen(object): pass
+
+class SetInitializer(InitializerBase):
+    __slots__ = ('_set','verified')
+
+    def __init__(self, init, allow_generators=True):
+        self.verified = False
+        if init is None:
+            self._set = None
+        else:
+            self._set = Initializer(
+                init, allow_generators=allow_generators,
+                treat_sequences_as_mappings=False)
+
+    def intersect(self, other):
+        if self._set is None:
+            if type(other) is SetInitializer:
+                self._set = other._set
+            else:
+                self._set = other
+        elif type(other) is SetInitializer:
+            if other._set is not None:
+                self._set = SetIntersectInitializer(self._set, other._set)
+        else:
+            self._set = SetIntersectInitializer(self._set, other)
+
+    def __call__(self, parent, idx):
+        if self._set is None:
+            return Any
+        else:
+            return self._set(parent, idx)
+
+    def constant(self):
+        return self._set is None or self._set.constant()
+
+    def setdefault(self, val):
+        if self._set is None:
+            self._set = ConstantInitializer(val)
+
+class SetIntersectInitializer(InitializerBase):
+    __slots__ = ('_A','_B',)
+    def __init__(self, setA, setB):
+        self._A = setA
+        self._B = setB
+
+    def __call__(self, parent, idx):
+        return SetIntersection(self._A(parent, idx), self._B(parent, idx))
+
+    def constant(self):
+        return self._A.constant() and self._B.constant()
+
+class RangeSetInitializer(InitializerBase):
+    __slots__ = ('_init', 'default_step',)
+    def __init__(self, init, default_step=1):
+        self._init = Initializer(init, treat_sequences_as_mappings=False)
+        self.default_step = default_step
+
+    def __call__(self, parent, idx):
+        val = self._init(parent, idx)
+        if not isinstance(val, collections.Sequence):
+            val = (1, val, self.default_step)
+        if len(val) < 3:
+            val = tuple(val) + (self.default_step,)
+        ans = RangeSet(*tuple(val))
+        ans.construct()
+        return ans
+
+    def constant(self):
+        return self._init.constant()
+
+    def setdefault(self, val):
+        # This is a real range set... there is no default to set
+        pass
 
 #
 # DESIGN NOTES
@@ -2636,8 +2347,8 @@ class Set(IndexedComponent):
         # HACK to make the "counted call" syntax work.  We wait until
         # after the base class is set up so that is_indexed() is
         # reliable.
-        if self._init_values.__class__ is _IndexedCallInitializer:
-            self._init_values = _CountedCallInitializer(self, self._init_values)
+        if self._init_values.__class__ is IndexedCallInitializer:
+            self._init_values = CountedCallInitializer(self, self._init_values)
 
     def construct(self, data=None):
         if self._constructed:
@@ -2653,7 +2364,7 @@ class Set(IndexedComponent):
             tmp_init, self._init_values = self._init_values, Initializer(
                     data, treat_sequences_as_mappings=False)
         try:
-            if type(self._init_values) is _ItemInitializer:
+            if type(self._init_values) is ItemInitializer:
                 for index in iterkeys(self._init_values._dict):
                     # The index is coming in externally; we need to
                     # validate it
