+++ pyomo/core/tests/unit/test_expr_misc.py
